+++ CHANGELOG.md
@@ -2,9 +2,14 @@ Race Protocol: A multi-chain infrastructure for asymmetric competitive games
 
 # Master(Unreleased)
 
+## Features
+- CLI: Add `claim` command to claim tokens from a recipient account.
+
+## Enhancements
+- SDK: Add caching for token/NFT fetching.
+
 ## Breaking changes
 - CONTRACT: Remove `claim_amount_cap`.
-- CLI: Add `claim` command to claim tokens from a recipient account.
 
 # 0.2.0